Update Prescriptions with the concept of a group:
A group denotes a logically similar set of prescription candidates
e.g. all Statins. We find this useful as it is an excellent concept
to run aggregations over from a non-domain expert POV.

The only way to seed this is editorially, thus there is a create_groups
management command which, as of right now, simply creates the statin
grouping, but could, in principle, be expanded.